restore: Move serviceWorker specific code into ServiceWorkerStore file
